ulong all the ints. Overkill, but Github IDs on `status` events are already more than a 32bit signed int can handle.

+++ Matterhook.NET/Webhooks/Discourse/Shared.cs
@@ -2,8 +2,8 @@
 {
     public class Actions_Summary
     {
-        public int id { get; set; }
-        public int count { get; set; }
+        public ulong id { get; set; }
+        public ulong count { get; set; }
         public bool hidden { get; set; }
         public bool can_act { get; set; }
     }
